+++ test.sh
@@ -11,7 +11,7 @@ if [ $# -gt 0 ]
 fi
 
 if [ "$1" == "-d" ]; then
-	flask --app webapp.py --debug run 
+	flask --app webapp.py --debug run --exclude-patterns *webapp.log*
 else
 	flask --app webapp.py run 
 fi