+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Change log
-Date: 15/01/2023 
+Date: 16/01/2023 
 Release: v2.4.1
 - changed dimensions of spots in world dx spost charts
 - managed empty data in data providers for charts
@@ -10,6 +10,7 @@ Release: v2.4.1
 - fixed issue #28
 - cut comments too long
 - charts layout changed
+- reduced javascripts
 
 ___
 Date: 01/01/2023