+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Change log
-Date: 16/01/2023 
+Date: 18/01/2023 
 Release: v2.4.1
 - changed dimensions of spots in world dx spost charts
 - managed empty data in data providers for charts